# openapi.yaml
openapi: 3.1.0
info:
  title: FabzClean Modern Backend (Option B)
  version: "1.0.0"
  description: |
    OpenAPI 3.1 specification for the FabzClean modernized backend (Flask + MySQL).
    - JWT auth for customers
    - Session cookie for admin
    - Worker token auth (bearer)
servers:
  - url: https://{{host}}
    description: Production (replace {{host}} with your domain or static IP)
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token (Authorization: Bearer <token>). Used by customer endpoints."
    workerToken:
      type: http
      scheme: bearer
      bearerFormat: "worker-token"
      description: "Worker device token (Authorization: Bearer <worker_token>)."
    adminSession:
      type: apiKey
      in: cookie
      name: session
      description: "Admin session cookie (set after /admin/login)."
  schemas:
    Error:
      type: object
      properties:
        msg:
          type: string
      example:
        msg: "Bad credentials"
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
        duration_minutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, price]
    ServiceCreate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        duration_minutes:
          type: integer
      required: [name, price]
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, email]
...
paths:
...
  /api/v1/customers/:
    get:
      summary: Get current customer's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update current customer's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
              example:
                name: "New Name"
                phone: "9999999999"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/services/:
    get:
      summary: List services
      responses:
        "200":
          description: Array of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      summary: Create a service (admin or privileged)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Validation/error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/services/{service_id}:
    parameters:
      - name: service_id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get service by id
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    put:
      summary: Update a service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      summary: Delete a service
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                example:
                  msg: "Deleted"
...
  /admin/api/customers:
    get:
      summary: List all customers (admin session required)
      security:
        - adminSession: []
      responses:
        "200":
          description: Array of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
        "401":
          description: Admin login required
...
security:
  - {}
...